+++ src/main/java/com/pharmquest/pharmquest/domain/post/web/controller/PostController.java
@@ -1,6 +1,5 @@
 package com.pharmquest.pharmquest.domain.post.web.controller;
 
-import com.google.protobuf.Api;
 import com.pharmquest.pharmquest.domain.mypage.domain.PostScrap;
 import com.pharmquest.pharmquest.domain.post.converter.PostCommentConverter;
 import com.pharmquest.pharmquest.domain.post.converter.PostConverter;
@@ -11,7 +10,7 @@ import com.pharmquest.pharmquest.domain.post.data.enums.Country;
 import com.pharmquest.pharmquest.domain.post.data.enums.PostCategory;
 import com.pharmquest.pharmquest.domain.post.data.mapping.Comment;
 import com.pharmquest.pharmquest.domain.post.data.mapping.PostLike;
-import com.pharmquest.pharmquest.domain.post.service.PostCommandService;
+import com.pharmquest.pharmquest.domain.post.service.post.PostCommandService;
 import com.pharmquest.pharmquest.domain.post.service.comment.PostCommentService;
 import com.pharmquest.pharmquest.domain.post.service.like.PostLikeService;
 import com.pharmquest.pharmquest.domain.post.service.scrap.PostScrapService;
@@ -24,13 +23,10 @@ import com.pharmquest.pharmquest.global.apiPayload.ApiResponse;
 import com.pharmquest.pharmquest.domain.token.JwtUtil;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.headers.Header;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
@@ -63,11 +59,11 @@ public class PostController {
 
     @GetMapping("/posts/{post_id}")
     @Operation(summary = "게시글 상세조회 API")
-    public ApiResponse<PostResponseDTO.PostDetailDTO> getPost(@Parameter (hidden = true) @RequestHeader("Authorization") String authorizationHeader,@PathVariable(name = "post_id")Long postId){
+    public ApiResponse<PostResponseDTO.PostDetailDTO> getPost(@Parameter (hidden = true) @RequestHeader("Authorization") String authorizationHeader,@PathVariable(name = "post_id")Long postId, @RequestParam(name="page")Integer page){
      
         User user = jwtUtil.getUserFromHeader(authorizationHeader);
 
-        PostResponseDTO.PostDetailDTO postDetail = postCommandService.getPost(user.getId(), postId);
+        PostResponseDTO.PostDetailDTO postDetail = postCommandService.getPost(user.getId(), postId,page);
 
         return ApiResponse.onSuccess(postDetail);
     }
